requirements/prod.txt
aiohttp[speedups]==3.8.1
fastapi[all]==0.78.0
factory-boy==3.2.1
matplotlib==3.5.2
mpld3==0.5.7
retry==0.9.2